oauth2.py
from jose import jwt, JWTError
from datetime import datetime, timedelta

SECRET_KEY = "8f0e552a31e2e8023389819b24508c5b3b13451f655be5b3e219ce0cfd330ca5"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.now() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    to_encode.update()

    encoded_jwt = jwt.encode(
        to_encode, SECRET_KEY, algorithm=ALGORITHM
    )
    return encoded_jwt
