requirements.txt
fastapi
uvicorn
black
pydantic
psycopg2-binary
python-dotenv
SQLAlchemy
email-validator
passlib[bcrypt]
python-jose